Move grading related code into existing module
Project info tags relating to grading are imported to empty view
The empty view is imported into the urls package to ensure that it is loaded at startup.
This is all necessary to register the tags at startup

The templates are all renamed only (with modifications re. renaming),
despite some being registered as deleted and added
