+++ bp/grading/orga/views.py
@@ -1,0 +1,2 @@
+# necessary to load the project info tags
+from . import project_info_tags
