+++ bp/templates/bp/grading/ag/project_grade.html
@@ -21,6 +21,6 @@
         </form>
     </turbo-frame>
 
-    {% include "bp/render_all_gradings.html" %}
+    {% include "bp/grading/ag/render_all_gradings.html" %}
 
 {% endblock %}
