+++ bp/templates/bp/grading/ag/project_grade_success.html
@@ -30,6 +30,6 @@
         </table>
     </turbo-frame>
 
-    {% include "bp/render_all_gradings.html" %}
+    {% include "bp/grading/ag/render_all_gradings.html" %}
 
 {% endblock %}
