+++ bp/templates/bp/grading/orga/project_info_misc_grading_content.html
@@ -1,0 +1,1 @@
+{% include "bp/grading/orga/render_gradings_tab.html" %}
