+++ bp/templates/bp/grading/orga/project_info_misc_grading_desc.html
