--- bp/templates/bp/project_info_misc_grading_content.html
@@ -1,1 +1,0 @@
-{% include "bp/render_gradings_tab.html" %}
